+++ app/jobs/match_feedback_email_job.rb
@@ -2,7 +2,7 @@ class MatchFeedbackEmailJob < ApplicationJob
   def perform(feedback_id, person)
     feedback = Feedback.find_by(id: feedback_id)
     if feedback.present?
-      UserMailer.with(user: person, feedback: feedback).match_feedback.deliver_now
+      ExpertMailer.with(expert: person, feedback: feedback).match_feedback.deliver_now
     end
   end
 end
