+++ spec/services/match_mailer_service_spec.rb
@@ -2,6 +2,7 @@
 
 require 'rails_helper'
 describe MatchMailerService do
+  let!(:national_referent) { create :user, :national_referent }
 
   before do
     ENV['APPLICATION_EMAIL'] = 'contact@mailrandom.fr'
